# data-web/openapi-section.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
...
                  "/github-repository":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "github-repository"
                                                                  ],
                                                         "summary":  "Lister les repositories GitHub",
                                                         "description":  "Cette opération permet de récupérer une liste paginée de repositories GitHub, avec recherche textuelle.",
                                                         "operationId":  "listGitHubRepositories",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "page",
                                                                                "in":  "query",
                                                                                "description":  "Index de page (commence à 0).",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  0
                                                                                           },
                                                                                "example":  0
                                                                            },
                                                                            {
                                                                                "name":  "size",
                                                                                "in":  "query",
                                                                                "description":  "Nombre d\u0027éléments par page.",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  10
                                                                                           },
                                                                                "example":  10
                                                                            },
                                                                            {
                                                                                "name":  "q",
                                                                                "in":  "query",
                                                                                "description":  "Texte libre de recherche (owner, name, fullName, URL, description, branche, langage).",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           },
                                                                                "example":  "spring"
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "400":  {
                                                                                       "description":  "Bad Request",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "string"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Not Found",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "string"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   },
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/GitHubRepositoryListResponse"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "post":  {
                                                          "tags":  [
                                                                       "github-repository"
                                                                   ],
                                                          "summary":  "Créer un repository GitHub",
                                                          "description":  "Cette opération permet de créer un nouveau repository GitHub dans la base applicative.",
                                                          "operationId":  "createGitHubRepository",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/GitHubRepositoryCreationForm"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "400":  {
                                                                                        "description":  "Bad Request",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "string"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "404":  {
                                                                                        "description":  "Not Found",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "string"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/GitHubRepository"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      },
                                             "delete":  {
                                                            "tags":  [
                                                                         "github-repository"
                                                                     ],
                                                            "summary":  "Supprimer tous les repositories GitHub",
                                                            "description":  "Cette opération permet de supprimer tous les repositories GitHub enregistrés.",
                                                            "operationId":  "deleteAllGitHubRepositories",
                                                            "responses":  {
                                                                              "400":  {
                                                                                          "description":  "Bad Request",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "string"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      },
                                                                              "404":  {
                                                                                          "description":  "Not Found",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "string"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      },
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          }
                                                        }
                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "GitHubRepository":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "id":  {
                                                                                              "type":  "string",
                                                                                              "description":  "Identifiant unique du repository.",
                                                                                              "example":  "123e4567-e89b-12d3-a456-426614174000"
                                                                                          },
                                                                                   "owner":  {
                                                                                                 "type":  "string",
                                                                                                 "description":  "Le propriétaire du repository.",
                                                                                                 "example":  "tiogars"
                                                                                             },
                                                                                   "name":  {
                                                                                                "type":  "string",
                                                                                                "description":  "Le nom court du repository.",
                                                                                                "example":  "data"
                                                                                            },
                                                                                   "fullName":  {
                                                                                                    "type":  "string",
                                                                                                    "description":  "Le nom complet owner/name du repository.",
                                                                                                    "example":  "tiogars/data"
                                                                                                },
                                                                                   "url":  {
                                                                                               "type":  "string",
                                                                                               "description":  "L\u0027URL du repository.",
                                                                                               "example":  "https://github.com/tiogars/data"
                                                                                           },
                                                                                   "description":  {
                                                                                                       "type":  "string",
                                                                                                       "description":  "Description fonctionnelle du repository.",
                                                                                                       "example":  "Application de gestion de données"
                                                                                                   },
                                                                                   "defaultBranch":  {
                                                                                                         "type":  "string",
                                                                                                         "description":  "Branche par défaut.",
                                                                                                         "example":  "main"
                                                                                                     },
                                                                                   "language":  {
                                                                                                    "type":  "string",
                                                                                                    "description":  "Langage principal.",
                                                                                                    "example":  "TypeScript"
                                                                                                },
                                                                                   "stars":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int32",
                                                                                                 "description":  "Nombre d\u0027étoiles.",
                                                                                                 "example":  42
                                                                                             },
                                                                                   "archived":  {
                                                                                                    "type":  "boolean",
                                                                                                    "description":  "Indique si le repository est archivé.",
                                                                                                    "example":  false
                                                                                                }
                                                                               }
                                                            },
...
                                       "GitHubRepositoryCreationForm":  {
                                                                            "type":  "object",
                                                                            "properties":  {
                                                                                               "owner":  {
                                                                                                             "type":  "string",
                                                                                                             "description":  "Le propriétaire du repository.",
                                                                                                             "example":  "tiogars"
                                                                                                         },
                                                                                               "name":  {
                                                                                                            "type":  "string",
                                                                                                            "description":  "Le nom du repository.",
                                                                                                            "example":  "data"
                                                                                                        },
                                                                                               "url":  {
                                                                                                           "type":  "string",
                                                                                                           "description":  "L\u0027URL du repository.",
                                                                                                           "example":  "https://github.com/tiogars/data"
                                                                                                       },
                                                                                               "description":  {
                                                                                                                   "type":  "string",
                                                                                                                   "description":  "Description fonctionnelle du repository.",
                                                                                                                   "example":  "Application de gestion de données"
                                                                                                               },
                                                                                               "defaultBranch":  {
                                                                                                                     "type":  "string",
                                                                                                                     "description":  "Branche par défaut.",
                                                                                                                     "example":  "main"
                                                                                                                 },
                                                                                               "language":  {
                                                                                                                "type":  "string",
                                                                                                                "description":  "Langage principal.",
                                                                                                                "example":  "TypeScript"
                                                                                                            },
                                                                                               "stars":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int32",
                                                                                                             "description":  "Nombre d\u0027étoiles.",
                                                                                                             "example":  42
                                                                                                         },
                                                                                               "archived":  {
                                                                                                                "type":  "boolean",
                                                                                                                "description":  "Indique si le repository est archivé.",
                                                                                                                "example":  false
                                                                                                            }
                                                                                           }
                                                                        },
...
                                       "GitHubRepositoryListResponse":  {
                                                                            "type":  "object",
                                                                            "properties":  {
                                                                                               "items":  {
                                                                                                             "type":  "array",
                                                                                                             "items":  {
                                                                                                                           "$ref":  "#/components/schemas/GitHubRepository"
                                                                                                                       }
                                                                                                         },
                                                                                               "count":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int32"
                                                                                                         },
                                                                                               "page":  {
                                                                                                            "type":  "integer",
                                                                                                            "format":  "int32"
                                                                                                        },
                                                                                               "size":  {
                                                                                                            "type":  "integer",
                                                                                                            "format":  "int32"
                                                                                                        },
                                                                                               "query":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                           }
                                                                        },
...
                                   }
                   }
}